[MOV] clipboard: move clipboard store to correct folder
Task: 6389205
