+++ tests/clipboard/clipboard_plugin.test.ts
@@ -19,9 +19,9 @@ import { markdownLink } from "../../src/helpers/misc";
 import { toZone, zoneToXc } from "../../src/helpers/zones";
 import { Model } from "../../src/model";
 import { featurePluginRegistry } from "../../src/plugins/plugin_registries";
-import { ClipboardStore, MAX_FILE_SIZE } from "../../src/plugins/ui_stateful/clipboard";
 import { clipboardHandlersRegistries } from "../../src/registries/clipboardHandlersRegistries";
 import { DependencyContainer } from "../../src/store_engine/dependency_container";
+import { ClipboardStore, MAX_FILE_SIZE } from "../../src/stores/clipboard_store";
 import { NotificationStore } from "../../src/stores/notification_store";
 import { ViewportsStore } from "../../src/stores/viewports_store";
 import { XMLString } from "../../src/types/xlsx";
