+++ tests/composer/composer_store.test.ts
@@ -6,8 +6,8 @@ import { DateTime, jsDateToRoundNumber } from "../../src/helpers/dates";
 import { getCanonicalSymbolName } from "../../src/helpers/misc";
 import { toZone } from "../../src/helpers/zones";
 import { Model } from "../../src/model";
-import { ClipboardStore } from "../../src/plugins/ui_stateful/clipboard";
 import { DependencyContainer } from "../../src/store_engine/dependency_container";
+import { ClipboardStore } from "../../src/stores/clipboard_store";
 import { HighlightStore } from "../../src/stores/highlight_store";
 import { NotificationStore } from "../../src/stores/notification_store";
 import { Store } from "../../src/types/store_engine";
