+++ tests/pivots/pivot_insert.test.ts
@@ -1,7 +1,7 @@
 import { Model } from "../../src";
 import { PIVOT_INSERT_TABLE_STYLE_ID } from "../../src/constants";
 import { toZone } from "../../src/helpers/zones";
-import { ClipboardStore } from "../../src/plugins/ui_stateful/clipboard";
+import { ClipboardStore } from "../../src/stores/clipboard_store";
 import { copy, insertPivot, paste } from "../test_helpers/commands_helpers";
 import { getCell, getCellText, getTable } from "../test_helpers/getters_helpers";
 import { createModelFromGrid, setGrid } from "../test_helpers/helpers";
