+++ tests/repeat_commands_plugin.test.ts
@@ -16,13 +16,13 @@ import {
 } from "../src/types/commands";
 
 import { CoreCommand, Dimension, UID } from "../src";
-import { ClipboardStore } from "../src/plugins/ui_stateful/clipboard";
 import {
   repeatCommandTransformRegistry,
   repeatCoreCommand,
   repeatLocalCommandTransformRegistry,
 } from "../src/registries/repeat_transform_registry";
 import { AutomaticSumStore } from "../src/stores/automatic_sum_store";
+import { ClipboardStore } from "../src/stores/clipboard_store";
 import {
   activateSheet,
   addEqualCf,
