+++ tests/table/table_autofill_plugin.test.ts
@@ -3,8 +3,8 @@ import { AutofillStore } from "../../src/components/autofill/autofill_store";
 import { TableAutofillStore } from "../../src/components/autofill/table_autofill_store";
 import { CellComposerStore } from "../../src/components/composer/composer/cell_composer_store";
 import { DEFAULT_TABLE_CONFIG } from "../../src/helpers/table_presets";
-import { ClipboardStore } from "../../src/plugins/ui_stateful/clipboard";
 import { DependencyContainer } from "../../src/store_engine/dependency_container";
+import { ClipboardStore } from "../../src/stores/clipboard_store";
 import {
   copy,
   createTable,
