+++ tests/table/tables_plugin.test.ts
@@ -43,7 +43,7 @@ import {
 import { DEFAULT_BORDER_DESC } from "../../src/constants";
 import { DEFAULT_TABLE_CONFIG, TABLE_PRESETS } from "../../src/helpers/table_presets";
 import { CellEvaluationPlugin } from "../../src/plugins/evaluation/cell_evaluation/cell_evaluation_plugin";
-import { ClipboardStore } from "../../src/plugins/ui_stateful/clipboard";
+import { ClipboardStore } from "../../src/stores/clipboard_store";
 import { TABLE_STYLE_ALL_RED } from "../test_helpers/constants";
 import { makeStore, makeStoreWithModel } from "../test_helpers/stores";
